requirements.txt
Django==5.1.3
djangorestframework==3.15.2
python-dotenv==1.0.1
psycopg2-binary==2.9.10
djangorestframework-simplejwt==5.3.1
drf-yasg==1.21.8
django-filter==24.3
coverage==7.6.8
pillow==11.1.0

